Report file size as unsigned long long to delegate

+++ SSZipArchive/SSZipArchive.h
@@ -46,7 +46,7 @@
 - (void)zipArchiveWillUnzipFileAtIndex:(NSInteger)fileIndex totalFiles:(NSInteger)totalFiles archivePath:(NSString *)archivePath fileInfo:(unz_file_info)fileInfo;
 - (void)zipArchiveDidUnzipFileAtIndex:(NSInteger)fileIndex totalFiles:(NSInteger)totalFiles archivePath:(NSString *)archivePath fileInfo:(unz_file_info)fileInfo;
 
-- (void)zipArchiveProgressEvent:(NSInteger)loaded total:(NSInteger)total;
+- (void)zipArchiveProgressEvent:(unsigned long long)loaded total:(unsigned long long)total;
 @end
 
 #endif /* _SSZIPARCHIVE_H */
